Able to view map based on user input.
