+++ app/src/main/java/com/example/giftly/AddEventScreen.java
@@ -2,6 +2,7 @@ package com.example.giftly;
 
 import static android.content.ContentValues.TAG;
 import static com.example.giftly.Giftly.client;
+import com.example.giftly.handler.IEvent;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -38,8 +39,6 @@ import com.example.giftly.handler.Event;
 import com.example.giftly.handler.User;
 import com.google.common.util.concurrent.FutureCallback;
 import com.google.common.util.concurrent.Futures;
-import com.mapbox.api.directions.v5.DirectionsCriteria;
-import com.mapbox.api.directions.v5.MapboxDirections;
 import com.mapbox.api.geocoding.v5.MapboxGeocoding;
 import com.mapbox.api.geocoding.v5.models.CarmenFeature;
 import com.mapbox.api.geocoding.v5.models.GeocodingResponse;
@@ -53,7 +52,6 @@ import retrofit2.Response;
 
 public class AddEventScreen extends AppCompatActivity implements AdapterView.OnItemSelectedListener {
     final Calendar myCalendar= Calendar.getInstance();
-
     public Button button_create_event, button_cancel_adding;
     public TextView textbox_eventName, textbox_eventDate, textbox_eventLocation;
     private SharedPreferences sharedPreferences;
@@ -67,47 +65,6 @@ public class AddEventScreen extends AppCompatActivity implements AdapterView.OnI
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_create_new_event);
 
-        // Testing Geocoding Functionality
-        String Address = "Wayne State University";
-        //String test = "https://api.mapbox.com/search/geocode/v6/forward?q="+Address;
-
-
-        MapboxGeocoding mapboxGeocoding = MapboxGeocoding.builder()
-                .accessToken(getString(R.string.mapbox_access_token))
-                .query(Address)
-                .build();
-
-        TextView tv1 = (TextView)findViewById(R.id.testAddress);
-
-        mapboxGeocoding.enqueueCall(new Callback<GeocodingResponse>() {
-            @Override
-            public void onResponse(Call<GeocodingResponse> call, Response<GeocodingResponse> response) {
-                List<CarmenFeature> results = response.body().features();
-
-
-                if (results.size() > 0) {
-
-                    // Log the first results Point.
-                    Point firstResultPoint = results.get(0).center();
-                    Log.d(TAG, "LONGITUDE: " + firstResultPoint.longitude());
-                    Log.d(TAG, "LATITUDE: " + firstResultPoint.latitude());
-                    String lon = String.valueOf(firstResultPoint.longitude());
-                    String lat = String.valueOf(firstResultPoint.longitude());
-                    tv1.setText(lon);
-
-                } else {
-
-                    // No result for your request were found.
-                    Log.d(TAG, "onResponse: No result found");
-
-                }
-            }
-
-            @Override
-            public void onFailure(Call<GeocodingResponse> call, Throwable t) {
-                t.printStackTrace();
-            }
-        });
 
         //Theme: Fetch the current color of the background
         sharedPreferences = getSharedPreferences("MyPrefs", Context.MODE_PRIVATE);
@@ -155,7 +112,7 @@ public class AddEventScreen extends AppCompatActivity implements AdapterView.OnI
                 eventMap.put("eventStartDate", eventDate);
                 eventMap.put("eventName", textbox_eventName.getText().toString());
                 eventMap.put("eventType", spinner.getSelectedItemPosition());
-                //eventMap.put("eventLocation",textbox_eventLocation.getText().toString());
+                eventMap.put("eventLocation",textbox_eventLocation.getText().toString());
 
 
                 Futures.addCallback(
@@ -180,7 +137,6 @@ public class AddEventScreen extends AppCompatActivity implements AdapterView.OnI
                         }, Giftly.service);
             }
         });
-
 
 
         //TODO UNCOMMENT SPINNER CATEGOREY ADDITIONS AS THEY ARE IMPLEMENTED
