+++ app/src/main/res/layout/activity_create_new_event.xml
@@ -256,18 +256,6 @@
         app:layout_constraintTop_toTopOf="@+id/event_type_selection"
         app:srcCompat="@mipmap/ic_dropdown_icon" />
 
-        <TextView
-            android:id="@+id/testAddress"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="TextView"
-            app:layout_constraintBottom_toTopOf="@+id/button_cancel"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.288"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/copy_link"
-            app:layout_constraintVertical_bias="0.548" />
-
 
     </androidx.constraintlayout.widget.ConstraintLayout>
